+++ lib/active_record/turntable/connection_proxy.rb
@@ -7,13 +7,13 @@ module ActiveRecord::Turntable
     # for expiring query cache
     CLEAR_CACHE_METHODS = [:update, :insert, :delete, :exec_insert, :exec_update, :exec_delete, :insert_many].freeze
 
-    attr_reader :klass
+    attr_reader :klass, :default_shard, :default_current_shard
     attr_writer :spec
 
     def initialize(klass, options = {})
       @klass = klass
-      @master_shard = MasterShard.new(klass)
-      @default_current_shard = @master_shard
+      @default_shard = DefaultShard.new(klass)
+      @default_current_shard = @default_shard
       @mixer = ActiveRecord::Turntable::Mixer.new(self)
     end
 
@@ -25,14 +25,16 @@ module ActiveRecord::Turntable
              :change_column, :change_column_default, :rename_column, :add_index,
              :remove_index, :initialize_schema_information,
              :dump_schema_information, :execute_ignore_duplicate,
-             :query_cache_enabled, to: :master_connection
+             :query_cache_enabled, to: :default_connection
 
     def cluster
       klass.turntable_cluster
     end
 
     def transaction(options = {}, &block)
-      connection.transaction(options, &block)
+      with_master {
+        connection.transaction(options, &block)
+      }
     end
 
     def cache
@@ -55,7 +57,7 @@ module ActiveRecord::Turntable
     end
 
     def enable_query_cache!
-      master_connection.enable_query_cache!
+      default_connection.enable_query_cache!
 
       klass.turntable_connections.each do |_k, v|
         v.connection.enable_query_cache!
@@ -63,7 +65,7 @@ module ActiveRecord::Turntable
     end
 
     def disable_query_cache!
-      master_connection.disable_query_cache!
+      default_connection.disable_query_cache!
 
       klass.turntable_connections.each do |_k, v|
         v.connection.disable_query_cache!
@@ -75,7 +77,7 @@ module ActiveRecord::Turntable
     end
 
     def clear_query_cache
-      master_connection.clear_query_cache
+      default_connection.clear_query_cache
 
       klass.turntable_connections.each do |_k, v|
         v.connection.clear_query_cache
@@ -107,7 +109,7 @@ module ActiveRecord::Turntable
     end
 
     def to_sql(arel, binds = [])
-      master.connection.to_sql(arel, binds)
+      default_connection.to_sql(arel, binds)
     end
 
     def shards
@@ -126,12 +128,8 @@ module ActiveRecord::Turntable
       fixed_shard_entry[object_id] = shard
     end
 
-    def master
-      @master_shard
-    end
-
-    def master_connection
-      master.connection
+    def default_connection
+      default_shard.connection
     end
 
     def current_shard
@@ -183,6 +181,22 @@ module ActiveRecord::Turntable
       end
     end
 
+    def with_master
+      old = cluster.slave_enabled?
+      cluster.set_slave_enabled(false)
+      yield
+    ensure
+      cluster.set_slave_enabled(old)
+    end
+
+    def with_slave
+      old = cluster.slave_enabled?
+      cluster.set_slave_enabled(true)
+      yield
+    ensure
+      cluster.set_slave_enabled(old)
+    end
+
     # Send queries to all shards in this cluster
     # @param [Boolean] continue_on_error when a shard raises error, ignore exception and continue
     def with_all(continue_on_error = false)
@@ -200,10 +214,10 @@ module ActiveRecord::Turntable
       end
     end
 
-    # Send queries to master connection and all shards in this cluster
+    # Send queries to default connection and all shards in this cluster
     # @param [Boolean] continue_on_error when a shard raises error, ignore exception and continue
-    def with_master_and_all(continue_on_error = false)
-      ([master] + cluster.shards).map do |shard|
+    def with_default_and_all(continue_on_error = false)
+      ([default_shard] + cluster.shards).map do |shard|
         begin
           with_shard(shard) {
             yield
@@ -217,8 +231,8 @@ module ActiveRecord::Turntable
       end
     end
 
-    def with_master(&block)
-      with_shard(master) do
+    def with_default_shard(&block)
+      with_shard(default_shard) do
         yield
       end
     end
@@ -228,13 +242,13 @@ module ActiveRecord::Turntable
 
     %w(columns columns_hash column_defaults primary_keys).each do |name|
       define_method(name.to_sym) do
-        master.connection_pool.send(name.to_sym)
+        default_shard.connection_pool.send(name.to_sym)
       end
     end
 
     %w(data_source_exists?).each do |name|
       define_method(name.to_sym) do |*args|
-        master.connection_pool.with_connection do |c|
+        default_shard.connection_pool.with_connection do |c|
           c.schema_cache.send(name.to_sym, *args)
         end
       end
@@ -242,26 +256,26 @@ module ActiveRecord::Turntable
 
     def columns(*args)
       if args.size > 0
-        master.connection_pool.columns[*args]
+        default_shard.connection_pool.columns[*args]
       else
-        master.connection_pool.columns
+        default_shard.connection_pool.columns
       end
     end
 
     def pk_and_sequence_for(*args)
-      master.connection.send(:pk_and_sequence_for, *args)
+      default_shard.connection.send(:pk_and_sequence_for, *args)
     end
 
     def primary_key(*args)
-      master.connection.send(:primary_key, *args)
+      default_shard.connection.send(:primary_key, *args)
     end
 
     def supports_views?(*args)
-      master.connection.send(:supports_views?, *args)
+      default_shard.connection.send(:supports_views?, *args)
     end
 
     def spec
-      @spec ||= master.connection_pool.spec
+      @spec ||= default_shard.connection_pool.spec
     end
 
     private