+++ lib/active_record/turntable/slave_shard.rb
@@ -1,0 +1,9 @@
+require "active_record/turntable/shard"
+
+module ActiveRecord::Turntable
+  class SlaveShard < Shard
+    def support_slave?
+      false
+    end
+  end
+end