+++ spec/config/turntable.rb
@@ -3,11 +3,11 @@ cluster :user_cluster do
 
   sequencer :user_seq, :mysql, connection: :user_seq
 
-  shard      1...20_000,     to: :user_shard_1
-  shard 20_000...40_000,     to: :user_shard_2
-  shard 40_000...60_000,     to: :user_shard_1
-  shard 60_000...80_000,     to: :user_shard_2
-  shard 80_000...10_000_000, to: :user_shard_3
+  shard      1...20_000,     to: :user_shard_1, slaves: [:user_shard_1_1]
+  shard 20_000...40_000,     to: :user_shard_2, slaves: [:user_shard_2_1]
+  shard 40_000...60_000,     to: :user_shard_1, slaves: [:user_shard_1_1]
+  shard 60_000...80_000,     to: :user_shard_2, slaves: [:user_shard_2_1]
+  shard 80_000...10_000_000, to: :user_shard_3, slaves: [:user_shard_3_1]
 end
 
 cluster :event_cluster do